+++ include/TrickHLA/ExecutionConfiguration.hh
@@ -37,6 +37,7 @@ NASA, Johnson Space Center\n
 #define TRICKHLA_EXECUTION_CONFIGURATION_HH
 
 // System include files.
+#include <cstdint>
 
 // TrickHLA include files.
 #include "TrickHLA/ExecutionConfigurationBase.hh"
@@ -60,8 +61,8 @@ class ExecutionConfiguration : public TrickHLA::ExecutionConfigurationBase
    friend void init_attrTrickHLA__ExecutionConfiguration();
 
   public:
-   double    run_duration;          ///< @trick_units{s}  The run duration of the simulation.
-   long long run_duration_microsec; ///< @trick_units{us} The run duration in microseconds.
+   double  run_duration;          ///< @trick_units{s}  The run duration of the simulation.
+   int64_t run_duration_microsec; ///< @trick_units{us} The run duration in microseconds.
 
    int   num_federates;      ///< @trick_units{--} Number of required federates.
    char *required_federates; ///< @trick_units{--} Comma-separated list of required federates.
