+++ source/TrickHLA/CTETimelineBase.cpp
@@ -39,7 +39,6 @@ NASA, Johnson Space Center\n
 */
 
 // System includes
-#include <cstdint>
 
 // Trick include files.
 
@@ -93,7 +92,7 @@ long long CTETimelineBase::wall_clock_time()
 {
    struct timespec tp;
    clock_gettime( clk_id, &tp );
-   return (int64_t)tp.tv_sec * 1000000LL + (int64_t)( ( tp.tv_nsec ) / 1000 );
+   return (long long)tp.tv_sec * 1000000LL + (long long)( ( tp.tv_nsec ) / 1000 );
 }
 
 /*!
